Check bincount in data coming from backend
The backend may return fewer data points than expected. This can e.g. be
observed for the hackrf_sweep backend on high system load. Currently
this leads to an error when trying to insert the data into the
HistoryBuffer and causes the application to crash:

ValueError: cannot copy sequence with size 24750 to array axis with dimension 25000

As a workaround, just ignore data in this case.